use a common jest config
